更新使用MySQL数据库，舍弃使用Oracle，因为报一些莫名其妙的错误

+++ src/main/java/com/alan/hysupermarket/service/ICategoryService.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 public interface ICategoryService {
 
-    List<Category> list();
+	List<Category> list();
 
     void add(Category category);
 
-    void delete(long id);
+    void delete(int id);
 
-    Category get(long id);
+    Category get(int id);
 
     void update(Category category);
