+++ src/main/java/com/alan/hysupermarket/service/IOrdersService.java
@@ -6,24 +6,24 @@ import com.alan.hysupermarket.pojo.OrdersItem;
 import java.util.List;
 
 public interface IOrdersService {
-    String waitPay = "waitPay";
-    String waitDelivery = "waitDelivery";
-    String waitConfirm = "waitConfirm";
-    String waitReview = "waitReview";
-    String finish = "finish";
-    String delete = "delete";
+	String waitPay = "waitPay";
+	String waitDelivery = "waitDelivery";
+	String waitConfirm = "waitConfirm";
+	String waitReview = "waitReview";
+	String finish = "finish";
+	String delete = "delete";
 
-    void add(Orders Orders);
+	void add(Orders Orders);
 
-    float add(Orders Orders, List<OrdersItem> ois);
+	float add(Orders Orders, List<OrdersItem> ois);
 
-    void delete(long id);
+	void delete(int id);
 
-    void update(Orders Orders);
+	void update(Orders Orders);
 
-    Orders get(long id);
+	Orders get(int id);
 
-    List<Orders> list();
+	List<Orders> list();
 
-    List<Orders> list(long uid, String excludedStatus);
+	List<Orders> list(int uid, String excludedStatus);
 }
